AUD-64: Refactor code
