+++ src/Application/AuditApplication.php
@@ -3,8 +3,8 @@
 namespace SetBased\Audit\Application;
 
 use SetBased\Audit\Command\AboutCommand;
-use SetBased\Audit\MySql\Command\DiffCommand;
 use SetBased\Audit\MySql\Command\AuditCommand;
+use SetBased\Audit\MySql\Command\DiffCommand;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Command\Command;
 
@@ -36,7 +36,8 @@ class AuditApplication extends Application
     $commands[] = new AboutCommand();
     $commands[] = new AuditCommand();
     $commands[] = new DiffCommand();
-    $commands[] = new ConfigCommand();
+
+//    $commands[] = new ConfigCommand();
 
     return $commands;
   }
