+++ src/MySql/Sql/AlterAuditTableAddColumns.php
@@ -2,8 +2,8 @@
 //----------------------------------------------------------------------------------------------------------------------
 namespace SetBased\Audit\MySql\Sql;
 
-use SetBased\Audit\MySql\Table\Columns;
-use SetBased\Audit\MySql\Table\ColumnType;
+use SetBased\Audit\MySql\Metadata\ColumnMetadata;
+use SetBased\Audit\MySql\Metadata\TableColumnsMetadata;
 use SetBased\Helper\CodeStore\MySqlCompoundSyntaxCodeStore;
 
 //----------------------------------------------------------------------------------------------------------------------
@@ -23,7 +23,7 @@ class AlterAuditTableAddColumns
   /**
    * The array of new columns for adding to table.
    *
-   * @var Columns
+   * @var TableColumnsMetadata
    */
   private $columns;
 
@@ -38,10 +38,10 @@ class AlterAuditTableAddColumns
   /**
    * Object constructor.
    *
-   * @param string  $auditSchemaName The name of the audit schema.
-   * @param string  $tableName       The name of the table.
-   * @param Columns $columns         The metadata of the new columns of the audit table (i.e. the audit columns and
-   *                                 columns of the data table).
+   * @param string               $auditSchemaName The name of the audit schema.
+   * @param string               $tableName       The name of the table.
+   * @param TableColumnsMetadata $columns         The metadata of the new columns of the audit table (i.e. the audit
+   *                                              columns and columns of the data table).
    */
   public function __construct($auditSchemaName, $tableName, $columns)
   {
@@ -61,7 +61,7 @@ class AlterAuditTableAddColumns
     $code = new MySqlCompoundSyntaxCodeStore();
 
     $code->append(sprintf('alter table `%s`.`%s`', $this->auditSchemaName, $this->tableName));
-    /** @var ColumnType $column */
+    /** @var ColumnMetadata $column */
     foreach ($this->columns->getColumns() as $column)
     {
       $code->append(sprintf('  add `%s` %s', $column->getProperty('column_name'), $column->getProperty('column_type')), false);
