+++ src/MySql/Sql/CreateAuditTrigger.php
@@ -2,8 +2,8 @@
 //----------------------------------------------------------------------------------------------------------------------
 namespace SetBased\Audit\MySql\Sql;
 
-use SetBased\Audit\MySql\Table\Columns;
-use SetBased\Audit\MySql\Table\ColumnType;
+use SetBased\Audit\MySql\Metadata\ColumnMetadata;
+use SetBased\Audit\MySql\Metadata\TableColumnsMetadata;
 use SetBased\Exception\FallenException;
 use SetBased\Exception\RuntimeException;
 use SetBased\Helper\CodeStore\MySqlCompoundSyntaxCodeStore;
@@ -26,7 +26,7 @@ class CreateAuditTrigger
   /**
    * AuditApplication columns from metadata.
    *
-   * @var Columns
+   * @var TableColumnsMetadata
    */
   private $auditColumns;
 
@@ -61,7 +61,7 @@ class CreateAuditTrigger
   /**
    * Audit columns from metadata.
    *
-   * @var Columns
+   * @var TableColumnsMetadata
    */
   private $tableColumns;
 
@@ -90,15 +90,15 @@ class CreateAuditTrigger
   /**
    * Creates a trigger on a table.
    *
-   * @param string   $dataSchemaName  The name of the data schema.
-   * @param string   $auditSchemaName The name of the audit schema.
-   * @param string   $tableName       The name of the table.
-   * @param string   $triggerAction   The trigger action (i.e. INSERT, UPDATE, or DELETE).
-   * @param string   $triggerName     The name of the trigger.
-   * @param Columns  $auditColumns    The audit table columns.
-   * @param Columns  $tableColumns    The data table columns.
-   * @param string   $skipVariable    The skip variable.
-   * @param string[] $additionalSql   Additional SQL statements.
+   * @param string               $dataSchemaName  The name of the data schema.
+   * @param string               $auditSchemaName The name of the audit schema.
+   * @param string               $tableName       The name of the table.
+   * @param string               $triggerAction   The trigger action (i.e. INSERT, UPDATE, or DELETE).
+   * @param string               $triggerName     The name of the trigger.
+   * @param TableColumnsMetadata $auditColumns    The audit table columns.
+   * @param TableColumnsMetadata $tableColumns    The data table columns.
+   * @param string               $skipVariable    The skip variable.
+   * @param string[]             $additionalSql   Additional SQL statements.
    */
   public function __construct($dataSchemaName,
                               $auditSchemaName,
@@ -193,7 +193,7 @@ class CreateAuditTrigger
     $columnNames = '';
 
     // First the audit columns.
-    /** @var ColumnType $column */
+    /** @var ColumnMetadata $column */
     foreach ($this->auditColumns->getColumns() as $column)
     {
       if ($columnNames) $columnNames .= ',';
@@ -201,7 +201,7 @@ class CreateAuditTrigger
     }
 
     // Second the audit columns.
-    /** @var ColumnType $column */
+    /** @var ColumnMetadata $column */
     foreach ($this->tableColumns->getColumns() as $column)
     {
       if ($columnNames) $columnNames .= ',';
@@ -222,10 +222,10 @@ class CreateAuditTrigger
     $values = '';
 
     // First the values for the audit columns.
-    /** @var ColumnType $column */
+    /** @var ColumnMetadata $column */
     foreach ($this->auditColumns->getColumns() as $column)
     {
-      $column = $column->getType();
+      $column = $column->getProperties();
       if ($values) $values .= ',';
       switch (true)
       {
@@ -255,7 +255,7 @@ class CreateAuditTrigger
     }
 
     // Second the values for the audit columns.
-    /** @var ColumnType $column */
+    /** @var ColumnMetadata $column */
     foreach ($this->tableColumns->getColumns() as $column)
     {
       if ($values) $values .= ',';
